Change Carousel View Model to HomeViewModel
